refactor(core): move `_DisplayHandler` to `recovery_device.layout`
Similar to `reset_device.layout` containing wordlist-related layouts.

[no changelog]
